Fix reap to continue returning deleted objects
With soft delete commit we started to use destroy in delete in all places. One
difference between delete and destroy is destroy does not support returning.
However we use returned child strands deleted by reap method.

With this commit, we are fixing reap behaviour to return deleted objects by
first materializing the query result and then deleting the records. We also
modified one of the tests to check the returned values.