Soft delete projects before archiving
For invoicing, we need to have project information, which would be lost when
users delete them, or more accuretely would be archived as DeletedRecords. We
don't want to refer to DeletedRecords during regular operation of the Clover.
With this commit, instead of archiving project entities, we soft delete them
by marking them as not visible. At some point, those invisible projects needs
to be cleaned up manually or we can write a cronjob.