Simplify Prog::Base.reap
Previously, we tried to replicate Sequel's returning.delete behaviour, which
uses DB.fetch in the background and returns rows instead of models. However,
there is only one piece of code in our codebase that depends on that particular
behaviour. By modifying depending piece of code to work with models, we can
significantly simplify the reap behaviour.