Update used hugepages when destroying a Vm.
We increase number of used hugepages when allocating a Vm. A Vm will
fail to allocate if there are not enough hugepages.

Previously we didn't update the number of hugepages when destroying
a Vm, which caused in not being able to allocate new Vms.

This PR fixes that by updating number of used hugepages when destroying
a Vm.