Update available storage size when destroying a Vm
We decrease vmhost available storage size when allocating a Vm. A Vm
will fail to allocate if there are not enough available storage.

Previously we didn't update the size of available storage when
destroying a Vm, which caused in not being able to allocate new Vms.

This PR fixes that by updating available storage size when destroying
a Vm.