Use YAML file for billing rates instead of a database table
Before this commit, billing rates were backed by a database table, because we
anticipated that the number of billing rates would be in the order of 10k as we
add new products and regions. However keeping this information in DB makes it
difficult to see current list of billing rates. Adding new billing raters also
requires database migration.

10k is high but still manageable as an in-memory array. We also have a quite a
long way until we reach that number. So, we are changing billing rates to be
backed by a YAML file.