Allocate VMs in a host with low memory/core ratio.
Previously we expected the host memory/core ratio be >= VMs
memory/core ratio. Now that we use 8GiB memory/core ratio,
allocating VMs in a an empty AX161 host failed, because it
had 4GiB memory/core ratio (32 cores and 128GiB memory).

But we could actually allocate VMs in that case by limiting
the number of cores that could be assigned to VMs instead.
For example in the AX161 case, instead of failing the allocation
we could limit ourselves only to 16 cores so we can still maintain
the 8GiB memory/core ratio.

This PR does that.