Remove SPI from policy creation
SPI is not really necessary in policies. For outgoing connections, once the
template is matched the correct SA is identified locally via reqid. For
incoming connections, if the template matches, spi is already in the header
of the packet, which is used to match the correct SA. Therefore, we do not need
spi in policy.
We also set the reqid for to 0 for the destination. That way, the incoming
packet can be matched flexibly with the correct SA. reqid 0 means, no reqid
preference.
When we implement rekeying, this change will be handy since we won't be dealing
with destination policies separately.