Increase test coverage of CloverBase
`freeze` override is in CloverVase so CloverWeb and CloverApi have this
new version. But we are freezing Clover not nested ones.

I moved this version of freeze to Clover

We should not call `DB.freeze` for tests too, because it blocks to
CloverWeb/CloverApi to enable new plugins that use new pg extensions.