Show selected location's private subnets while creating vm
We need to decide to show which private subnets dynamically depending on
selected location. So it requires JavaScript code. We need to pass list
of subnets for that location to frontend code. There are two ways: pass
statically when page is rendered or pass dynamically via an ajax HTTP
request. Latter one requires an extra endpoint on backend to return
subnets. I implemented first option. If number of subnets increased too
much for each location we can consider fetching dynamically for selected
location.

Safari doesn't allow hiding select options, because of that we are
disabling them.