Fix rekeying nic synchronization issue
This commit introduces new properties to keep the refresh_keys logic in
sync with actually rekeying nic strands. The problem was that since we
iterate over nics in multiple states, it must make sure the snapshot of
the nics taken at the first state of rekeying does match in later
stages. That's why we introduce a function named rekeying_nics and
filter the related nic objects appropriately.