Remove refresh_mesh logic
With this commit, we are removing the logic to refresh_mesh. There are
3 main reasons to do that;
1. Current refresh_mesh is completely on dataplain instances. This makes
it hard to develop the code iteratively, because we have to always think
about the migration of the scripts.
2. Current refresh_mesh destroys all the state and policy objects and
recreates them, this causes connection drop in the whole mesh. We don't
want that.
3. There is a huge code overlap with the refresh_keys logic. State
creation, policy creation/update. We can just reuse the code.

The logic is mainly depending on the synchronization of NicNexus and
SubnetNexus strands. They are 2 separate entities, there are many nic
objects in a private subnet and we have to make sure a tunnel's two end
get into the rekeying sequence together. If not, that tunnel wouldn't
work because of the nature of the ipsec tunnelling. More information
regarding that is provided in the commit 85420cc. Therefore, we
implement a pingpong logic in between subnet and nic nexus' to manage
when to create new state objects and update policies. That is managed
via semaphores.

There is a known errata here. Using strand labels directly here to
pick which entities to work on doesn't work well with "push" of
sub progs. In future, we can fix this by adding a fit to purpose
column to the nic entity. For now, this is good enough.