Adds an explanation to hetzner api for not fetching additional IPv6s
Recently we realized that when we added an IPv6 block (additional, by mistake),
the VmHost.create_addresses did not create the ipv6 block entity in the
database. Looking into the pull_ips, I realized that the ipv6 blocks come with
active_server_ip as the main ipv6 address of the server, not the ipv4. Since the
find_matching_ips function filters the ip addresses according to the host
address in sshable, this was expected. It's harmless for now but we should
tackle this once we start implementing extra ipv6 block support. That's why the
comment is added to show guidance for future.