Start billing when/if only vm is accessible to the customer
Before this commit, we created billing records way early in the nexus. With this
commit, we move them to the run state. That way, if a provisioning fails
in earlier stages, we won't immediately start billing the customer.
There was also a consideration to use wait_sshable but we wanted to
avoid that since in future, we may just have VMs without ssh access.