Add active and finalize helpers to BillingRecord
An active billing record implies that its span does not have an upper
bound. We perform this condition check in several locations. With the
`active` dataset helper, we can execute this check with ease.

When we wish to conclude a billing record, we update its upper limit
with `Time.now`. We can encapsulate this operation using the `finalize`
helper.