Block some ips in private subnet to be assigned
We should not allow the first ip in the private subnet to be assigned to
a resource. This is necessary because the first ip in the subnet is used
for routing in the host. If we assign it to a VM, too, then it will not
be accessible because the packets will be accepted by the nc* interface
and not be sent to the vm.
We are also blocking the next 3 IPs after the network address for future
proofing. We might need them in future for different purposes.
We further block the very last ip because typically, the last ip is used
as a broadcasting ip. AWS also does this and the information can be seen
here https://docs.aws.amazon.com/vpc/latest/userguide/subnet-sizing.html