Cap number of billable hours in month to 672
We use 1 month = 672 hours for conversions. Number of hours in a month changes
between 672 and 744, We are also capping number billable hours in month to 672
while generating invoices. This ensures that users won't see higher price in
their invoice compared to price calculator and also we charge same amount no
matter the number of days in a given month.