Improve BootstrapRhizome's idempotency
It is possible that setup step of BootstrapRhizome fails and we want to
retry it. A retry would have failed before because user "rhizome" and
its home directory had been created before.

This PR improves that by deleting the user and its home directory.

Note that `userdel` will fail (even if with the `-rf` args) if the user
doesn't exist, hence the `|| true` to ignore its failure.

This was all I had to do to make this retriable if the last `echo`
statement failed. To make sure if this is enough changes to ensure
idempotency, we need to look more.