Add a HetznerAPI method to get the main IP4 address.
This is handy for end-to-end CI as it will require less configuration.

The existing pull_ips requires a `VmHost` to exist, and requires some
work to fetch the main IP4 address out of its results, so adding this
simpler method which has less requirements and fetches only the main IP4
address.