Use cp reflink to copy boot image for unencrypted disks.
Previously we always used spdk_dd to copy disk images to storage disk
files. For encrypted case there are no other easy ways, as data needs to
be encrypted in the disk file.

But for unencrypted disks, we could alternatively use Linux's `cp`
utility with the `--reflink=auto` arg. This will make copying on file
systems that support reflink much faster. It will just do a shallow copy
and will copy blocks on write.

To make use of the speed-up, before setting up a VmHost:

* Make sure you have a spare disk which you can format as btrfs
* `mkfs.btrfs /dev/disk-device-name`
* `mkdir -p /var/storage`
* Add necessary fstab entries to mount the device to /var/storage, and
  also do the mount manually.

Then install VmHost and you should be good to go.

Note that this PR also changes the image path to `/var/storage/images`.