Start using Kea for DHCP4
With this commit, we are moving the DHCP4 functionality from Dnsmasq to
Kea. Main reason to do so is to that Dnsmasq doesn't support prefix
delegation for DHCP6, so we decided that we will use Kea for DHCP in
general. This is the first step. In this commit, we also prepare host
with the related packages. Therefore, the DHCP6 path will simply
implement the dhcp6 configuration and daemon.