Add and use "husk" process to diagnose slow termination issues
`husk` can be a catch-all place to put process supervision code.

Right now, we're interested in reviewing threads that are still
running briefly after a SIGTERM is sent.  This is done by sending a
subsequent SIGQUIT after a little while.

To try it out, do something like this, in two terminals.  One should run:

    $ ./bin/husk ruby -e "Signal.trap('SIGTERM') { }; sleep 86400"

The other can run:

    pkill -SIGTERM -f husk

In the first terminal, you can check the exit status:

    $ echo $?
    131

This is the expected exit status for a SIGQUIT:

    clover-development(main)> Signal.list["QUIT"] + 128
    => 131