Combine rhizome Gemfiles
Installing multiple Gemfiles to their own separate directories does not work if
they need to be used together, which is the case when you call functions defined
in common from other rhizome folders. It does not work because  installed gems
are searched at bundle installation path relative to working directory. When
gems are installed to separate directories, it cannot find some of them.

Thus, gems need to be installed to same directory, which can cause conflicts or
overwrites. To prevent such problems, gems need to be managed from one Gemfile,
which is what this commit does. We create rhizome/Gemfile, which globs Gemfiles
in the child directories and combines them.