Allow to download an image from a custom URL
The assumption is that images without links already exist on the VM
host. I manually downloaded the "github-ubuntu-2204" image to the hosts.

We handle runner images differently because GitHub cannot publicly share
these images due to legal issues
(https://github.com/actions/runner-images/issues/176), and we must
respect the same restrictions. Additionally, the image size is 86GB.

Instead of using a hardcoded image link, we provide our download URL at
runtime.

The "download-boot-image" executable serves as an interface for
downloading an image with or without a custom URL. The creation of an
operational prog for it will be my next step.