Print ubids when paging
It's more terse, and easier to double-click these to select, plus the
first two bytes can convey useful type information.

There is a small downside: previously, `Strand[uuid]` was closer to an
expression to *always* get a strand value, and often, the first thing
I want to do with a deadline violation is `.run`.

But seeing that selection with the `-[]` characters is hard, and
what's printed isn't even a valid expression anyway (no quotes), let's
move to UBID and I think there are other ways to streamline getting
the Strand value.