Remove all YYY and TODO from the code base
At least one can be converted into a comment, but most of these are
unremarkable good intentions, no more interesting than the many other
good intentions that are not written.

Some are so terse, the bring to mind Fermat's quotation: "for this, I
have found a truly wonderful proof, but the margin is too small to
contain it."

TODOs are fine for, like, a prototype, where they are to be to-done,
and in some way supervised via text search to form a list of
things...to do.