Use pg short code for PostgresServer
Currently pg is used for Page model, but it is an internal model and doesn't
need a very good representation compared to PostgresServer, which is much more
visible and the primary model of a key service. We decided to bite the bullet
and make pg the short code for PostgresServer while it is still early.