Use separate pipes for apoptosis and completion notification
Previous versions of this patch had two pairs of pipes, but I
momentarily tried to collapse them into one.  The problem is
synchronizing `close` in relation to `IO.select`, or ignoring it given
our use case.

Unfortunately, running `IO.select` on closed pipes yields an
`IOError`, which is a general exception type, and it doesn't seem to
have a good field to test for this precise situation, unlike some
other places where we might have been able to rescue a specific
`Errno` exception.

Given that, paying for two pairs of pipes so the `close` resource
management is simpler seems to make more sense.