Don't try to purge_storage if storage hasn't been created yet.
If a VM is destroyed early on before anything has been created, then
storage artifacts won't exist, and there's nothing there to be deleted.

We got to know of this issue when we get an exception which tried to
read `prep.json` in `purge_storage` and it didn't exist. In which case
if everything was done in correct order, storage root shouldn't also
exist, and this patch handles that case.

On the other hand, if `prep.json` doesn't exist and storage root exists,
that's an exceptional case and we should still raise an error.