Do not raise an exception while destroying a private subnet has active nics
A private subnet cannot be deleted if it has active nics. Currently, it
raises an exception in such cases. However, raising exceptions doesn't
offer any benefits and only clutters the logs with stack traces.
Additionally, when a deletion attempt fails, it immediately retries.
Daniel has implemented a backoff mechanism (#607), but let's reserve it
for handling unexpected errors.

This situation is expected when both the private subnet's destroy
semaphore and the virtual machine's destroy semaphore are increased.
Eventually, the virtual machine will be deleted, followed by the
deletion of the private subnet.

In addition, printing a log message and adding a 5-second nap reduce
the load on respirate. Otherwise, it would repeatedly run
'PrivateSubnetNexus' until the virtual machine is deleted.

I registered a page for it; if it gets stuck indefinitely, we'll receive
a page.

The recent increase in these logs is due to my actions. When destroying
a GitHub runner, I simultaneously increase the destroy semaphore for
both the subnet and the vm. I cannot wait for the vm's deletion before
destroying the private subnet because the runner can only access the
private subnet through the vm. Once the vm is deleted, the private
subnet record becomes inaccessible from the runner