Add destroy label to the vm host nexus
Recently, we decommissioned existing vm hosts and reconfigured them as
GitHub runner vm hosts. I handled the deletion of the vm hosts manually.
However, it's beneficial to automate the condition checks.  It changes
the allocation state and then waits. Since allocation can occur
simultaneously with the updating of the 'allocation_state' bit, it's
good to wait for 5 seconds to ensure that no virtual machine is assigned
to this host during the process. Subsequently, it waits to confirm that
no virtual machine host exists before proceeding with the destruction.