Replace daemonizer with systemd-run on GitHub runners
rhizome common includes a useful script called 'daemonizer' for running
non-blocking long-running commands. I've utilized it for running GitHub
Actions runner scripts as well. To use it, we need to install rhizome on
the runner.

However, it's worth noting that the entire runner provisioning process
takes approximately 1 minute, while the bootstrap rhizome step alone
takes 20 seconds. Installing rhizome solely for the purpose of using
'daemonizer' may not be the most optimal approach.

daemonizer offers a wide range of features, but the runner doesn't
require all of them. Therefore, we have decided to replace daemonizer
with 'systemd-run', which is also used under the hood of daemonizer