Prevent race conditions in adduser
In production, we hit multiple cases of group being created but user did not.
After closer inspection, we realized that adduser runs few subcommands like
useradd and groupadd. If there are concurrent adduser commands running, it is
possible that the first useradd and second groupadd command pick same id due
to race condition, thus causing adduser command to fail in midway and leave
half created users in the system. When there are half created users in the
system, next run of create_unix_user fails with "adduser: The group `xxxxxx'
already exists" and stucks.

As a solution, we assign random id between 1100 and 59999. Ids smaller than
1000 and larger than 59999 are soft-reserved by OS. We also reserve 100 ids
for our own users such as rhizome and spdk. If randomly generated id is in
use already, which has low probability because we wouldn't create more than
few hundred users on one host even in the case of shared cores, the command
would fail and we would retry adding user with a different id in next run.