Add a helper function to retrieve the subject of the strand
There is a one-to-one correspondence between a strand and its subject,
typically sharing the same ID. The strand class has a NAVIGATE list for
accessing associated objects. Prog base also provides a 'subject_is'
helper, which the prog utilizes to derive the subject. The NAVIGATE list
proves particularly beneficial for interacting with strands, especially
when they don't load prog or lack the 'subject_is' helper.
We have more classes than those listed as strand subjects in NAVIGATE.

Given our ability to identify the source model of the UBID, we can use
the subject helper to retrieve it, eliminating the need to add all
models to the NAVIGATE list.