Pin postgres version to 15.4
PostgreSQL 16 has been released, and it includes some changes regarding
the bootstrap user. Our CI using the latest tag, so it's using
postgres:16 currently.  It  is failing with the following error.

    ERROR:  permission denied to alter role
    DETAIL:  The bootstrap user must have the SUPERUSER attribute.

We are using "postgres 15.4" at development and production.

I pin CI version to 15.4.

When we want to upgrade postgres 16, we may change all of them.