Generate invoice for a single project
Currently, InvoiceGenerator produces invoices for all projects within a
given timestamp. I need to generate an invoice for a single project to
display the current usage to the customer.

I plan to generate an invoice for a single project, covering the period
from the last invoice to today, to demonstrate current usage.