Change the invoice filename format
The invoice filenames are crucial for those who collect invoices. We can
generate invoice files with more desirable names for accounting purposes
with minimal effort. We want the invoice names to include a sortable
date and be fully qualified. Additionally, they should have a unique id
to search easily.

The new format is `Ubicloud-<YEAR>-<MONTH>-<INVOICE_NUMBER>.pdf`.

The PDF view derives the filename from the page's title. Currently, we
append the `Ubicloud -` prefix to all titles. However, it contains
spaces, which are undesirable in filenames. Therefore, if `@full_page`
is true, we omit this prefix and use the invoice filename as the title.