Return the list of invoice models from the InvoiceGenerator
Currently, `InvoiceGenerator` returns a list of invoice content hashes.
If `save_result` is set to `true`, it also persists this data to the
database. The persisted invoice data is then serialized using
`Serializers::Web::Invoice` for display in the UI.

The `Invoice` model offers several advantages over simple content
hashes. It allows us to utilize model helpers and serializers.  We don't
need to persist data to access the invoice model; we can construct it
without saving.

My plan is to generate an invoice for a single project's current usage
and display it using `Serializers::Web::Invoice` with the current
template files.

However, several changes are required to return a non-persisted invoice
model:
-  Before saving to the database, the keys in `Invoice.content` are
   symbols. However, upon retrieval from the database, these keys become
   strings. `Serializers::Web::Invoice` expects these keys to be strings
   as it uses the invoice data from the database. I used `content:
   JSON.parse(project_content.to_json)` to convert symbol keys to
   strings.
-  I also converted keys to strings in the `check_invoice_for_single_vm`
   helper.
-  When we store a numeric value in a JSONB column, it's persisted in
   scientific notation, which isn't very useful in JSON data. We already
   convert numeric cost and amount to float when setting them in the
   invoice content. As `credit` is also a numeric value, we need to cast
   it to float as well.