Add created_at column to Vm and VmHost
The "created_at" field is a useful property for various reasons. It's
particularly beneficial during debugging. Moreover, in the event of
breaking changes, we can modify behavior based on the age of the
resource, or decide which resources to migrate based on their creation
date.

Additionally, we don't display virtual machines in any specific order.
The order is arbitrary, determined by PostgreSQL. However, I have now
changed it to display from the newest to the oldest, top to bottom.