Add a helper to redeliver failed GitHub webhook deliveries
GitHub does not resend webhook deliveries that have failed. It is
necessary to inspect the failed deliveries and redeliver them in case of
missed deliveries due to deployments or incidents.

The "/app/hook/deliveries" endpoint lacks a filter, necessitating
pagination until the specified date. It also returns all attempts for
each delivery packet. Each delivery packet has a unique "guid".

We group deliveries by "guid". If there isn't a successful delivery
within a "guid" group, we initiate redelivery for that particular
packet.

    # Get failed deliveries for last 10 minutes
    Github.failed_deliveries(Time.now - 10 * 60)

    # Redeliver failed deliveries for last 10 minutes
    Github.redeliver_failed_deliveries(Time.now - 10 * 60)

Since it paginates until the specified date is reached, avoid providing
a very old date as it could result in an excessive number of pagination
requests.

GitHub also recommends to use such a script to redeliver failed
deliveries

https://docs.github.com/en/webhooks/using-webhooks/automatically-redelivering-failed-deliveries-for-a-github-app-webhook