Add the current usage to UI
At present, we display only finalized invoices, which prevents customers
from knowing their current usage until the end of the month. Therefore,
it would be beneficial to display the current usage on the billing page.

I generate an invoice with `save_result: false` from the end of the last
invoice of a project to now. If the project doesn't have an invoice yet,
I use the first day of the month. If a project lacks any billing
records, `InvoiceGenerator` returns an empty list. Consequently, I
create a new invoice at `project.current_invoice`.

Since the current invoice isn't persisted to the database, it doesn't
have an `ubid`. For invoice details, `/invoice/current` takes precedence
over `/invoice/{ubid}`.

I moved the credit and discount information to the project's
billing page.