Implement a prog to download a boot image to the vm host
Under normal circumstances, the boot image is downloaded during the
provisioning process if the vm host doesn't already have it. This isn't
usually a problem for smaller images, for instance, ubuntu-jammy, which
is 2.2GB, only slightly delays the initial provisioning on the vm host.

However, GitHub runner image is considerably larger at 86GB. The process
of downloading and converting it takes approximately 30 minutes, so
it's essential to download this image prior to the first provisioning.

Furthermore, GitHub releases new image versions weekly, necessitating
updates to our image every week. Therefore, it's beneficial to have a
prog that can manage the entire operation.

The vm host cannot accept new vm allocations during image updates
because we delete the old image and download the new one. As mentioned,
this process takes around 30 minutes for GitHub Actions image, so any
runner allocation on that vm host will likely attempt to download the
image again and possibly time out.